01/05 Done Done after comments !!!

+++ src/animals/Bear.java
@@ -18,13 +18,17 @@ public class Bear extends Animal {
         loadImages("bea");
         cor_x3 = -size/2;
         cor_x4 = 0;
-        cor_y1 = (int) (-30-size/5);
+        cor_y1 = -30-size/5;
         cor_y3 = (int) (size*0.3);
         cor_x5 = -size*6/7;
         cor_y5 = cor_y6 = -size/3;
-        cor_h = (int)(size*2/3);
+        cor_h = size*2/3;
     }
 
+
+    /**
+     * @return the type of the bear
+     */
     @Override
     public EFoodType getFoodtype() {return EFoodType.MEAT;}
